fix: resolve pre-existing -Werror build warnings and add command wrapper functions
- Created cmd_wrappers.c with args_to_argv helper and wrapper functions for
  13 commands using incompatible int (*)(int, char**) signature
- Updated cmd_table.inc to use wrapper functions, fixing 13 type mismatch errors
- Added wrapper declarations to shell_cmds.h
- Fixed tmpfs.c: added missing void *priv parameter to all VFS operations
- Fixed syntax.c: removed unused in_target and expected_close variables
- Fixed net.c loopback_send parameter type to avoid overflow warning
- Fixed net.h: updated net_loopback_send declaration to match
- Fixed net_udp.c: corrected format specifiers and unused parameter
- Fixed wireguard.c: removed unused variable, added unused attribute,
  fixed shift overflow, initialized poly_key before use
- Fixed service.c: corrected format specifiers for int64_t

+++ src/include/net.h
@@ -215,7 +215,7 @@ void ipv6_send_rs(void);
 
 /* Loopback interface */
 int  net_loopback_init(void);
-int  net_loopback_send(const void *data, uint16_t len);
+int  net_loopback_send(const void *data, int len);
 
 /* TCP keepalive support */
 void net_tcp_set_keepalive(int conn_id, int keepalive);
